initial project state, first draft
