--- autofac/101/Autofac.101/App/Bootstrapper.cs
@@ -1,10 +1,0 @@
-﻿namespace Autofac101.App
-{
-	public class Bootstrapper
-	{
-		public static Application Bootstrap()
-		{
-
-		}
-	}
-}
