--- autofac/101/Autofac.101/Program.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using Autofac101.App;
-
-namespace Autofac101
-{
-	public class Program
-	{
-		public static void Main(string[] args)
-		{
-			var app = Bootstrapper.Bootstrap();
-			app.Run();
-			Console.WriteLine("Ready to exit");
-			Console.ReadLine();
-		}
-	}
-}
