--- autofac/101/Autofac.101/Services/ProvideManners/FrenchMannersProvider.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace Autofac101.Services.ProvideManners
-{
-	public class FrenchMannersProvider : IProvideWelcome, IProvideFarewell
-	{
-		public void Greet()
-		{
-			Console.WriteLine("Bonjour!");
-		}
-
-		public void Farewell()
-		{
-			Console.WriteLine("Au Revoir!");
-		}
-	}
-}
