--- autofac/101/Autofac.101/Services/ProvideManners/IProvideGreeting.cs
@@ -1,7 +1,0 @@
-﻿namespace Autofac101.Services.ProvideManners
-{
-	public interface IProvideWelcome
-	{
-		void Greet();
-	}
-}
