--- autofac/101/Autofac.101/Services/ProvideManners/SpanishMannersProvider.cs
@@ -1,17 +1,0 @@
-﻿using System;
-
-namespace Autofac101.Services.ProvideManners
-{
-	public class SpanishMannersProvider : IProvideWelcome, IProvideFarewell
-	{
-		public void Greet()
-		{
-			Console.WriteLine("¡Hola!");
-		}
-
-		public void Farewell()
-		{
-			Console.WriteLine("¡Adiós!");
-		}
-	}
-}
