--- autofac/101/Autofac101.sln
@@ -1,22 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25123.0
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Autofac101", "Autofac.101\Autofac101.csproj", "{3ACA7489-7CA2-4787-A62A-E50C2AA5901E}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3ACA7489-7CA2-4787-A62A-E50C2AA5901E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3ACA7489-7CA2-4787-A62A-E50C2AA5901E}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3ACA7489-7CA2-4787-A62A-E50C2AA5901E}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3ACA7489-7CA2-4787-A62A-E50C2AA5901E}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
