+++ autofac/Autofac101/Autofac101.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Autofac101", "Autofac101\Autofac101.csproj", "{0E15233C-A124-4BAC-A6EF-5392BC0A0AEF}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{0E15233C-A124-4BAC-A6EF-5392BC0A0AEF}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{0E15233C-A124-4BAC-A6EF-5392BC0A0AEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{0E15233C-A124-4BAC-A6EF-5392BC0A0AEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{0E15233C-A124-4BAC-A6EF-5392BC0A0AEF}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
