+++ autofac/Autofac101/Autofac101/App/Bootstrapper.cs
@@ -1,0 +1,11 @@
+﻿using Autofac101.Services.ProvideManners;
+
+namespace Autofac101.App
+{
+	public class Bootstrapper
+	{
+		public static Application Bootstrap()
+		{
+		}
+	}
+}
