+++ autofac/Autofac101/Autofac101/Program.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using Autofac101.App;
+
+namespace Autofac101
+{
+	public class Program
+	{
+		public static void Main(string[] args)
+		{
+			var app = Bootstrapper.Bootstrap();
+			app.Run();
+			Console.WriteLine("Program completed successfully");
+			Console.ReadLine();
+		}
+	}
+}
