+++ autofac/Autofac101/Autofac101/Services/ProvideManners/EnglishMannersProvider.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Autofac101.Services.ProvideManners
+{
+	public class EnglishMannersProvider : IProvideWelcome, IProvideFarewell
+	{
+		public void Greet()
+		{
+			Console.WriteLine("Hello!");
+		}
+
+		public void Farewell()
+		{
+			Console.WriteLine("Goodbye!");
+		}
+	}
+}
