+++ autofac/Autofac101/Autofac101/Services/ProvideManners/FrenchMannersProvider.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Autofac101.Services.ProvideManners
+{
+	public class FrenchMannersProvider : IProvideWelcome, IProvideFarewell
+	{
+		public void Greet()
+		{
+			Console.WriteLine("Bonjour!");
+		}
+
+		public void Farewell()
+		{
+			Console.WriteLine("Au Revoir!");
+		}
+	}
+}
