+++ autofac/Autofac101/Autofac101/Services/ProvideManners/IProvideFarewell.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Autofac101.Services.ProvideManners
+{
+	public interface IProvideFarewell
+	{
+		void Farewell();
+	}
+}
