+++ autofac/Autofac101/Autofac101/Services/ProvideManners/IProvideGreeting.cs
@@ -1,0 +1,7 @@
+﻿namespace Autofac101.Services.ProvideManners
+{
+	public interface IProvideWelcome
+	{
+		void Greet();
+	}
+}
