+++ autofac/Autofac101/Autofac101/Services/ProvideManners/SpanishMannersProvider.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace Autofac101.Services.ProvideManners
+{
+	public class SpanishMannersProvider : IProvideWelcome, IProvideFarewell
+	{
+		public void Greet()
+		{
+			Console.WriteLine("¡Hola!");
+		}
+
+		public void Farewell()
+		{
+			Console.WriteLine("¡Adiós!");
+		}
+	}
+}
